+++ src/core/constants.h
@@ -4,7 +4,7 @@ namespace adhan {
 
 constexpr const char* kAppName = "Adhan Volume";
 constexpr const char* kAppFolderName = "AdhanVolume";
-constexpr const char* kVersion = "1.0.0";
+constexpr const char* kVersion = "1.0.2";
 constexpr const char* kMutexName = "Local\\AdhanVolumeSingleton";
 
 constexpr int kConfigVersion = 1;
@@ -14,9 +14,15 @@ constexpr int kScheduleVersion = 1;
 constexpr int DEFAULT_ADHAN_DURATION_SECONDS = 180;
 constexpr int DEFAULT_FADE_DURATION_MS = 4000;
 constexpr int DEFAULT_THRESHOLD_SECONDS = 60;
-constexpr int kSchedulerIdleIntervalMs = 20000;
+constexpr int kSchedulerIdleIntervalMs = 15000;
+constexpr int kSchedulerSoonIntervalMs = 5000;
+constexpr int kSchedulerNearIntervalMs = 1000;
+constexpr int kSchedulerSoonWindowMs = 600000;
+constexpr int kSchedulerNearWindowMs = 120000;
+constexpr int kClockJumpThresholdMs = 45000;
 constexpr int kFadeTickMs = 100;
 constexpr int kMuteHoldTickMs = 1000;
+constexpr float kVolumeVerifyEpsilon = 0.03f;
 constexpr int kCacheRetentionDays = 14;
 constexpr int kHttpTimeoutMs = 15000;
 constexpr int kHttpMaxRetries = 3;