+++ src/win/win_platform.h
@@ -6,8 +6,10 @@
 
 namespace adhan {
 
+class Logger;
+
 HttpClient* create_win_http_client();
-VolumeController* create_win_volume_controller();
+VolumeController* create_win_volume_controller(Logger* log = 0);
 std::string win_appdata_root();
 void win_enable_dpi_aware();